pyproject.toml
[tool.poetry]
name = "qa-guru-lesson-15"
version = "0.1.0"
description = "no description"
authors = ["Irina.Voron <yarutich.irina@gmail.com>"]
license = "no license"
readme = "README.md"
packages = [{include = "qa"}]

[tool.poetry.dependencies]
python = "^3.12"
selene = "2.0.0rc9"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
